+++ src/config.rs
@@ -34,3 +34,14 @@ impl Config {
         Config { listen , log_path: None}
     }
 }
+
+#[cfg(test)]
+mod test {
+    use crate::notification::Notification;
+
+    #[test]
+    fn test_countdown_char() -> () {
+        let notification = Notification::info("Hello".to_string());
+        notification.countdown_char();
+    }
+}